extract getJson into file
