+++ src/utilities/getJson.ts
@@ -1,0 +1,22 @@
+import getTokenJson from './getTokenJson'
+import buildFigmaData from './buildFigmaData'
+import UserSettings from '../../types/settings'
+/**
+ * @name getJson
+ * @param {PluginAPI} figma
+ * @param {boolean} stringify
+ */
+const getJson = (figma: PluginAPI, userSettings: UserSettings, stringify: boolean = true) => {
+  // construct figma data object
+  const figmaData = buildFigmaData(figma, {
+    prefix: userSettings.prefix,
+    excludePrefix: userSettings.excludePrefix
+  })
+  if (stringify === false) {
+    return getTokenJson(figmaData, 'styleDictionary', userSettings.nameConversion)
+  }
+  // get tokens as stringified json
+  return JSON.stringify(getTokenJson(figmaData, 'styleDictionary', userSettings.nameConversion))
+}
+
+export default getJson
